not 16

+++ Eigen 2014/validate_akin_test.py
@@ -20,7 +20,7 @@ def predict(model_path, input_directory, output_directory):
         f = np.load(input_directory)
         images_np = f['images']
         print(images_np.dtype)
-        f32_images_np = images_np.astype('float16')
+        f32_images_np = images_np.astype('float32')
         print(f32_images_np.dtype)
 
 
@@ -30,13 +30,13 @@ def predict(model_path, input_directory, output_directory):
         
         f32_images_np = tf.image.resize_images(f32_images_np, (height, width))
         print('\n** Loaded ' + str(f32_images_np.shape) + ' images. ** \n')
-        images = tf.convert_to_tensor(f32_images_np, dtype=tf.float16)
+        images = tf.convert_to_tensor(f32_images_np, dtype=tf.float32)
         images = tf.expand_dims(images,0)
 
         print('\n ** ' + str(tf.shape(images))+' ** \n')
 
         # Create a placeholder for the input image
-        input_node = tf.placeholder(dtype=tf.float16)
+        input_node = tf.placeholder(dtype=tf.float32)
         keep_conv = input_node
         keep_hidden = input_node
